requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
docutils==0.22.3
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
